Added about page, account information
Fixed the code to where it will now show the about page when you click on Video Game Library & Sales link in the nav bar when logged in. Fixed the code to where it will show the users account information when they click on account information. Changed the default picture for the account information.