Made updates to several pages and the tab
I added a favicon to the base and base2 templates to show the vgls logo in the tab. Added the copyright logo to the bottom left of the page stating that VGLS is a subsidiary of L.L.A.M.A. Inc. Adjusted the appearance of the register account page and the Change pasword page.