Fixed being able to delete posts, and added new images to possibly use for background wallpaper.
Can view the reviews on both the logged in side as well as the non logged in side. Only the logged in side can make adjustments to their reviews and delete them. Only the creator of the post can delete the post at this time. I have also exported an updated copy of my sql server database with a few reviews there